split safe tests from TestStorageContainerManagerHA
